refactor(b20_security): extract business logic from dispatch to token
Move security-specific methods from dispatch.rs to token.rs:
- Authorization helpers: ensure_security_operator, ensure_default_admin, ensure_burn_from_role
- Policy operations: is_supported_policy_scope, ensure_supported_policy_type, policy_id_checked, update_policy
- Share operations: to_shares, shares_of, update_share_ratio
- Security redeem: security_redeem, security_redeem_with_memo, security_redeem_burn, emit_redeemed
- Batch operations: batch_mint, batch_burn
- Minimum redeemable: update_minimum_redeemable
- Security identifiers: update_security_identifier

dispatch.rs now delegates to token.rs methods, keeping dispatch lean (routing only).
announce() stays in dispatch.rs since it requires inner_with_privilege for self-dispatch.
